wip: Remove Apollo from client/server
