+++ client/src/components/AssetsPage/AssetCard.tsx
@@ -10,7 +10,7 @@ import { NonFungibleToken } from '../../generated/graphql_schema';
 import { urlFromCid } from '../../api/ipfsUploader';
 import AssetTransfer from './AssetTransfer';
 import { useContracts } from '../App/globalContext';
-import useSettings from '../common/useSettings';
+import config from '../../config';
 
 const { Paragraph, Text, Link } = Typography;
 
@@ -64,7 +64,6 @@ interface Props {
 const AssetCard: FC<Props> = ({ token, onChange }) => {
   const [transferVisible, setTransferVisible] = useState(false);
   const contracts = useContracts();
-  const { settings } = useSettings();
 
   const handleTransfer = () => {
     if (!contracts) {
@@ -142,7 +141,7 @@ const AssetCard: FC<Props> = ({ token, onChange }) => {
                   Contract:{' '}
                   <Link
                     title={token.contractInfo.address}
-                    href={`${settings?.bcdGuiUrl}/${settings?.bcdNetwork}/${token.contractInfo.address}`}
+                    href={`${config.bcdGuiUrl}/${config.bcdNetwork}/${token.contractInfo.address}`}
                     target="_blank"
                   >
                     {token.contractInfo.name} - {token.contractInfo.address}
@@ -182,7 +181,7 @@ const AssetCard: FC<Props> = ({ token, onChange }) => {
           <Col span={12} css={{ display: 'flex', alignItems: 'center' }}>
             <Link
               css={{ margin: 'auto' }}
-              href={`${settings?.bcdGuiUrl}/${settings?.bcdNetwork}/${token.contractInfo.address}`}
+              href={`${config.bcdGuiUrl}/${config.bcdNetwork}/${token.contractInfo.address}`}
               target="_blank"
             >
               DETAILS
