+++ client/src/components/AssetsPage/Assets.tsx
@@ -4,7 +4,7 @@ import { jsx } from '@emotion/core';
 import { Row, Col, Spin } from 'antd';
 import { Zoom } from 'react-awesome-reveal';
 
-import { useNftsQuery } from '../common/useNftsQuery';
+import { useNftsQuery } from '../../hooks/useNftsQuery';
 import AssetCard from './AssetCard';
 import ContractsFilter from './ContractsFilter';
 import ContractsTitle from './ContractsTitle';
