+++ client/src/config.ts
@@ -1,0 +1,12 @@
+// TODO: Update with REACT_APP_ENV variables
+
+export default {
+  rpc: 'https://delphinet-tezos.giganode.io',
+  bcdApiUrl: 'https://api.better-call.dev',
+  bcdGuiUrl: 'https://better-call.dev',
+  bcdNetwork: 'delphinet',
+  contracts: {
+    nftFaucet: 'KT1CV1zsMnrnVHHZLKGdZ16bpSKhXfbJ8G8T',
+    nftFactory: 'KT1HDdJAXVxi3scXydzhLZio2KipbyPJSru5'
+  }
+};
