+++ client/src/resolvers/indexerStats.ts
@@ -1,0 +1,6 @@
+import { mkBetterCallDev } from './betterCallDev';
+
+export async function indexerStats(bcdApiUrl: string, bcdNetwork: string) {
+  const s = await mkBetterCallDev(bcdApiUrl, bcdNetwork).stats();
+  return s[0];
+}
