+++ server/src/index.ts
@@ -1,8 +1,54 @@
-import { importSchema } from 'graphql-import';
+import express, { Express, Response } from 'express';
+import IpfsClient from 'ipfs-http-client';
+import url from 'url';
+import bodyParser from 'body-parser';
+import fileUpload from 'express-fileupload';
+import http from 'http';
 
-import createContext from './components/context';
-import runHttpServer from './components/http_server';
-import path from 'path';
+// TODO: Move this configuration to a JSON definition
+const ipfsConfig = {
+  // The URL of our IPFS API server, our Web UI uploads files to.
+  apiUrl: 'http://ipfs:5001',
+
+  // The URL of our IPFS gateway server, which can be used for fast file download
+  // It is the same server as the one running IPFS API.
+  gatewayUrl: 'http://127.0.0.1:8080/',
+
+  // The URL of a public IPFS read-only gateway server. It may take time to
+  // propagate information from our IPFS server to a public one.
+  // It can be also used for file download but it may be very slow.
+  publicGatewayUrl: 'https://ipfs.io/'
+};
+
+function createHttpServer(app: Express) {
+  app.use(bodyParser.urlencoded({ extended: true }));
+  app.use(bodyParser.json());
+  app.use(
+    fileUpload({
+      limits: { fileSize: 10 * 1024 * 1024 } // 20MB
+    })
+  );
+
+  app.post('/ipfs-upload', async (req: any, res: Response) => {
+    if (!req.files?.file?.data) {
+      throw Error('No file data found');
+    }
+
+    const ipfsClient = IpfsClient(ipfsConfig.apiUrl);
+    const ipfsFile = await ipfsClient.add(req.files.file.data);
+    const cid = ipfsFile.cid.toString();
+
+    return res.status(200).send({
+      cid,
+      size: ipfsFile.size,
+      url: url.resolve(ipfsConfig.gatewayUrl, `ipfs/${cid}`),
+      publicGatewayUrl: url.resolve(ipfsConfig.publicGatewayUrl, `ipfs/${cid}`)
+    });
+  });
+
+  const httpServer = http.createServer(app);
+  return httpServer;
+}
 
 process.on('unhandledRejection', (reason, _promise) => {
   console.log('[Process] Unhandled Promise Rejection:', reason);
@@ -12,23 +58,16 @@ process.on('uncaughtException', error => {
   console.log('[Process] Uncaught Exception:', error);
 });
 
-const IS_DEV_ENV = !['production', 'test'].includes(process.env.NODE_ENV || '');
-
-const typeDefs = importSchema(path.resolve(__dirname + '/schema.graphql'));
-
 async function run() {
   const envPort = process.env.MINTER_API_PORT;
   const port = envPort ? parseInt(envPort) : 3300;
-  const context = await createContext();
-
-  runHttpServer(context, typeDefs, port, IS_DEV_ENV);
+  const app = express();
+  createHttpServer(app).listen(port, () => {
+    console.log(`[Server] Serving on port ${port}`);
+  });
 }
 
 async function main() {
-  // Uncomment below to view the environment on startup
-  // if (IS_DEV_ENV) {
-  //   console.log(process.env);
-  // }
   try {
     await run();
   } catch (e) {
